Working
Imported packer and useRef

added ref's for all the inputs so i can get their values.

File paths system is the same

Get the wordDoc then you have to do some formatting to make it saveable.

Blob to buffer to uint8Array. Shouldn't need to do this but doc package
doesn't work.

Write binary file is the only way to save uint8array with some error
catching code

Added the ref's to the elements in the dom.